requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.10
numpy==2.2.4
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pyngrok==7.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
slack_bolt==1.23.0
slack_sdk==3.35.0
sqlparse==0.5.3
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
